Updated test_core for dataset load testing
