+++ tests/test_core/test_datasets/convo_goldens.csv
@@ -1,0 +1,16 @@
+scenario,turns,expected_outcome,user_description,context
+User asks for weather forecast for Paris,user: What's the weather like in Paris this weekend? [retrieval_context: Weather API docs]|assistant: It's expected to be rainy with temperatures around 18°C. [retrieval_context: Weather API response],User receives accurate weather forecast,User is planning a trip to Paris and wants to check the weather,Weather|Paris
+User asks for a reminder to take medicine,user: Can you remind me to take my medicine at 8 PM tonight? [retrieval_context: Reminder service API]|assistant: Got it! I’ll remind you to take your medicine at 8 PM tonight. [retrieval_context: Reminder set confirmation],User receives a reminder to take their medicine,User needs a reminder to take their medication at a specified time,Health|Medication
+User translates a phrase from English to Japanese,user: How do you say 'thank you' in Japanese? [retrieval_context: Translation database]|assistant: You can say 'ありがとう' (arigatou).,Assistant correctly translates the phrase,User needs a quick translation,Translation|English to Japanese
+User books a restaurant reservation,user: Can you book a table for 2 at 7 PM tonight at Luigi's? [retrieval_context: Reservation service API]|assistant: Done! Your table for 2 at Luigi's is confirmed for 7 PM. [retrieval_context: Booking confirmation],Reservation is confirmed,User wants to book a dinner reservation for 2,Restaurants|Booking
+User asks for the latest sports scores,user: What's the score of the Liverpool vs Manchester United game? [retrieval_context: Sports API docs]|assistant: The score is Liverpool 2 - 1 Manchester United. [retrieval_context: Sports API response],User receives updated sports scores,User wants to check the latest score of a football match,Sports|Football
+User asks for a movie recommendation,user: Can you recommend a good movie to watch tonight? [retrieval_context: Movie recommendation service]|assistant: How about 'The Shawshank Redemption'? It's a classic!,User receives a relevant movie suggestion,User wants a movie suggestion for the weekend,Entertainment|Movies
+User checks their bank account balance,"user: What's my bank balance? [retrieval_context: Banking API]|assistant: Your current balance is $3,250.75. [retrieval_context: Bank account data]",User receives their current bank balance,User needs to check how much money they have in their bank account,Finance|Banking
+User checks the flight status for their flight,user: Is my flight to New York on time? [retrieval_context: Flight status API]|assistant: Your flight to New York is on time and will depart at 10:30 AM. [retrieval_context: Flight status API response],User receives accurate flight status,User is checking whether their flight is on time,Travel|Flights
+User asks for a health tip,user: Can you give me a health tip? [retrieval_context: Health database]|assistant: Make sure to drink plenty of water throughout the day and get at least 30 minutes of exercise. [retrieval_context: Health tips database],User receives a healthy lifestyle suggestion,User wants to know how to improve their health,Health|Wellness
+User asks for a nearby restaurant recommendation,"user: Where can I find a good sushi place near me? [retrieval_context: Restaurant API]|assistant: I recommend 'Sushi Hana', just a 5-minute walk from your location. [retrieval_context: Restaurant database]",User gets a restaurant suggestion,User wants a restaurant recommendation near their current location,Food|Restaurant
+User checks the weather for a different city,user: What's the weather like in Tokyo? [retrieval_context: Weather API]|assistant: The weather in Tokyo is sunny with a high of 25°C. [retrieval_context: Weather API response],User receives accurate weather data for a new city,User is checking the weather for a different location,Weather|Weather Forecast
+User asks for help with setting an alarm,user: Can you set an alarm for 7 AM tomorrow? [retrieval_context: Alarm service API]|assistant: Your alarm is set for 7 AM tomorrow. [retrieval_context: Alarm set confirmation],User's alarm is set successfully,User wants to set an alarm for 7 AM,Productivity|Alarm
+User asks for the time in a different timezone,user: What time is it in New York right now? [retrieval_context: Time API]|assistant: The current time in New York is 2:30 PM. [retrieval_context: Time API response],User receives the correct time in the specified timezone,User wants to know the current time in New York,Time|Timezone
+User asks for a nearby gym recommendation,"user: Where's the nearest gym? [retrieval_context: Gym database]|assistant: The nearest gym is 'FitZone', just a 3-minute walk from here. [retrieval_context: Gym database]",User receives a suggestion for a nearby gym,User is looking for a gym near them,Fitness|Gym
+User asks for a quick recipe,"user: Can you give me a quick recipe for dinner? [retrieval_context: Recipe database]|assistant: How about a quick pasta with tomato sauce? Just boil the pasta, heat some tomato sauce, and mix them together! [retrieval_context: Recipe suggestions]",User receives a simple recipe,User is looking for an easy recipe for dinner,Cooking|Recipes
