+++ tests/test_core/test_datasets/convo_goldens.json
@@ -1,0 +1,152 @@
+[
+    {
+        "scenario": "User asks for weather forecast for Paris",
+        "turns": "user: What's the weather like in Paris this weekend? [retrieval_context: Weather API docs]|assistant: It's expected to be rainy with temperatures around 18°C. [retrieval_context: Weather API response]",
+        "expected_outcome": "User receives accurate weather forecast",
+        "user_description": "User is planning a trip to Paris and wants to check the weather",
+        "context": [
+            "Weather",
+            "Paris"
+        ]
+    },
+    {
+        "scenario": "User asks for a reminder to take medicine",
+        "turns": "user: Can you remind me to take my medicine at 8 PM tonight? [retrieval_context: Reminder service API]|assistant: Got it! I’ll remind you to take your medicine at 8 PM tonight. [retrieval_context: Reminder set confirmation]",
+        "expected_outcome": "User receives a reminder to take their medicine",
+        "user_description": "User needs a reminder to take their medication at a specified time",
+        "context": [
+            "Health",
+            "Medication"
+        ]
+    },
+    {
+        "scenario": "User translates a phrase from English to Japanese",
+        "turns": "user: How do you say 'thank you' in Japanese? [retrieval_context: Translation database]|assistant: You can say 'ありがとう' (arigatou).",
+        "expected_outcome": "Assistant correctly translates the phrase",
+        "user_description": "User needs a quick translation",
+        "context": [
+            "Translation",
+            "English to Japanese"
+        ]
+    },
+    {
+        "scenario": "User books a restaurant reservation",
+        "turns": "user: Can you book a table for 2 at 7 PM tonight at Luigi's? [retrieval_context: Reservation service API]|assistant: Done! Your table for 2 at Luigi's is confirmed for 7 PM. [retrieval_context: Booking confirmation]",
+        "expected_outcome": "Reservation is confirmed",
+        "user_description": "User wants to book a dinner reservation for 2",
+        "context": [
+            "Restaurants",
+            "Booking"
+        ]
+    },
+    {
+        "scenario": "User asks for the latest sports scores",
+        "turns": "user: What's the score of the Liverpool vs Manchester United game? [retrieval_context: Sports API docs]|assistant: The score is Liverpool 2 - 1 Manchester United. [retrieval_context: Sports API response]",
+        "expected_outcome": "User receives updated sports scores",
+        "user_description": "User wants to check the latest score of a football match",
+        "context": [
+            "Sports",
+            "Football"
+        ]
+    },
+    {
+        "scenario": "User asks for a movie recommendation",
+        "turns": "user: Can you recommend a good movie to watch tonight? [retrieval_context: Movie recommendation service]|assistant: How about 'The Shawshank Redemption'? It's a classic!",
+        "expected_outcome": "User receives a relevant movie suggestion",
+        "user_description": "User wants a movie suggestion for the weekend",
+        "context": [
+            "Entertainment",
+            "Movies"
+        ]
+    },
+    {
+        "scenario": "User checks their bank account balance",
+        "turns": "user: What's my bank balance? [retrieval_context: Banking API]|assistant: Your current balance is $3,250.75. [retrieval_context: Bank account data]",
+        "expected_outcome": "User receives their current bank balance",
+        "user_description": "User needs to check how much money they have in their bank account",
+        "context": [
+            "Finance",
+            "Banking"
+        ]
+    },
+    {
+        "scenario": "User checks the flight status for their flight",
+        "turns": "user: Is my flight to New York on time? [retrieval_context: Flight status API]|assistant: Your flight to New York is on time and will depart at 10:30 AM. [retrieval_context: Flight status API response]",
+        "expected_outcome": "User receives accurate flight status",
+        "user_description": "User is checking whether their flight is on time",
+        "context": [
+            "Travel",
+            "Flights"
+        ]
+    },
+    {
+        "scenario": "User asks for a health tip",
+        "turns": "user: Can you give me a health tip? [retrieval_context: Health database]|assistant: Make sure to drink plenty of water throughout the day and get at least 30 minutes of exercise. [retrieval_context: Health tips database]",
+        "expected_outcome": "User receives a healthy lifestyle suggestion",
+        "user_description": "User wants to know how to improve their health",
+        "context": [
+            "Health",
+            "Wellness"
+        ]
+    },
+    {
+        "scenario": "User asks for a nearby restaurant recommendation",
+        "turns": "user: Where can I find a good sushi place near me? [retrieval_context: Restaurant API]|assistant: I recommend 'Sushi Hana', just a 5-minute walk from your location. [retrieval_context: Restaurant database]",
+        "expected_outcome": "User gets a restaurant suggestion",
+        "user_description": "User wants a restaurant recommendation near their current location",
+        "context": [
+            "Food",
+            "Restaurant"
+        ]
+    },
+    {
+        "scenario": "User checks the weather for a different city",
+        "turns": "user: What's the weather like in Tokyo? [retrieval_context: Weather API]|assistant: The weather in Tokyo is sunny with a high of 25°C. [retrieval_context: Weather API response]",
+        "expected_outcome": "User receives accurate weather data for a new city",
+        "user_description": "User is checking the weather for a different location",
+        "context": [
+            "Weather",
+            "Weather Forecast"
+        ]
+    },
+    {
+        "scenario": "User asks for help with setting an alarm",
+        "turns": "user: Can you set an alarm for 7 AM tomorrow? [retrieval_context: Alarm service API]|assistant: Your alarm is set for 7 AM tomorrow. [retrieval_context: Alarm set confirmation]",
+        "expected_outcome": "User's alarm is set successfully",
+        "user_description": "User wants to set an alarm for 7 AM",
+        "context": [
+            "Productivity",
+            "Alarm"
+        ]
+    },
+    {
+        "scenario": "User asks for the time in a different timezone",
+        "turns": "user: What time is it in New York right now? [retrieval_context: Time API]|assistant: The current time in New York is 2:30 PM. [retrieval_context: Time API response]",
+        "expected_outcome": "User receives the correct time in the specified timezone",
+        "user_description": "User wants to know the current time in New York",
+        "context": [
+            "Time",
+            "Timezone"
+        ]
+    },
+    {
+        "scenario": "User asks for a nearby gym recommendation",
+        "turns": "user: Where's the nearest gym? [retrieval_context: Gym database]|assistant: The nearest gym is 'FitZone', just a 3-minute walk from here. [retrieval_context: Gym database]",
+        "expected_outcome": "User receives a suggestion for a nearby gym",
+        "user_description": "User is looking for a gym near them",
+        "context": [
+            "Fitness",
+            "Gym"
+        ]
+    },
+    {
+        "scenario": "User asks for a quick recipe",
+        "turns": "user: Can you give me a quick recipe for dinner? [retrieval_context: Recipe database]|assistant: How about a quick pasta with tomato sauce? Just boil the pasta, heat some tomato sauce, and mix them together! [retrieval_context: Recipe suggestions]",
+        "expected_outcome": "User receives a simple recipe",
+        "user_description": "User is looking for an easy recipe for dinner",
+        "context": [
+            "Cooking",
+            "Recipes"
+        ]
+    }
+]
