+++ tests/test_core/test_datasets/goldens.csv
@@ -1,0 +1,16 @@
+input,actual_output,expected_output,retrieval_context,context,source_file
+What is the tallest mountain in the world?,Mount Everest,Mount Everest,Mountain heights list,Geography|Mountains,
+Solve 5 * 7,35,35,Basic arithmetic,Math|Multiplication,
+Define 'photosynthesis',Process by which green plants convert light energy into chemical energy,Process by which green plants convert light energy into chemical energy,Biology textbooks|Science API,Biology|Photosynthesis,
+What is the capital of France?,Paris,Paris,World capitals list,Geography|Capital Cities,
+What is the square root of 64?,8,8,Basic arithmetic,Math|Square Roots,
+What causes rain?,Rain is caused by the condensation of water vapor in the atmosphere.,Rain is caused by the condensation of water vapor in the atmosphere.,Weather patterns data,Science|Meteorology,
+Translate 'Hello' into Spanish.,Hola,Hola,Translation data,Language|Translation,
+What is the chemical formula of water?,H2O,H2O,Chemical compounds data,Chemistry|Molecules,
+What is the speed of light?,"299,792,458 meters per second","299,792,458 meters per second",Physics constants list,Physics|Speed of Light,
+Who wrote 'Romeo and Juliet'?,William Shakespeare,William Shakespeare,Literary works list,Literature|Shakespeare,
+What is the largest planet in our solar system?,Jupiter,Jupiter,Solar system data,Astronomy|Planets,
+How many continents are there?,7,7,Continents list,Geography|Continents,
+What is the boiling point of water?,100°C,100°C,Temperature data,Science|Temperature,
+What is 15% of 200?,30,30,Basic arithmetic,Math|Percentages,
+Who wrote 'Romeo and Juliet'?,William Shakespeare,William Shakespeare,Literary works list,Literature|Shakespeare,
